move out of analysis dir
