--- compiler/src/iree/compiler/Codegen/Common/GPU/Analysis/CMakeLists.txt
@@ -1,37 +1,0 @@
-################################################################################
-# Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# compiler/src/iree/compiler/Codegen/Common/GPU/Analysis/BUILD.bazel           #
-#                                                                              #
-# Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
-# CMake-only content.                                                          #
-#                                                                              #
-# To disable autogeneration for this file entirely, delete this header.        #
-################################################################################
-
-iree_add_all_subdirs()
-
-iree_cc_library(
-  NAME
-    Analysis
-  HDRS
-    "ThreadUniformAnalysis.h"
-  SRCS
-    "ThreadUniformAnalysis.cpp"
-  DEPS
-    LLVMSupport
-    MLIRAffineDialect
-    MLIRAnalysis
-    MLIRArithDialect
-    MLIRGPUDialect
-    MLIRIR
-    MLIRSCFDialect
-    MLIRSupport
-    iree::compiler::Codegen::Dialect::Codegen::IR::IREECodegenDialect
-    iree::compiler::Codegen::Dialect::GPU::IR::IREEGPUDialect
-    iree::compiler::Dialect::HAL::IR
-    iree::compiler::Dialect::TensorExt::IR
-    iree::compiler::Dialect::Util::IR
-  PUBLIC
-)
-
-### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
