+++ compiler/src/iree/compiler/Codegen/Common/GPU/BUILD.bazel
@@ -211,3 +211,28 @@ iree_tablegen_doc(
         "@llvm-project//mlir:PassBaseTdFiles",
     ],
 )
+
+iree_compiler_cc_library(
+    name = "Analysis",
+    srcs = [
+        "ThreadUniformAnalysis.cpp",
+    ],
+    hdrs = [
+        "ThreadUniformAnalysis.h",
+    ],
+    deps = [
+        "//compiler/src/iree/compiler/Codegen/Dialect/Codegen/IR:IREECodegenDialect",
+        "//compiler/src/iree/compiler/Codegen/Dialect/GPU/IR:IREEGPUDialect",
+        "//compiler/src/iree/compiler/Dialect/HAL/IR",
+        "//compiler/src/iree/compiler/Dialect/TensorExt/IR",
+        "//compiler/src/iree/compiler/Dialect/Util/IR",
+        "@llvm-project//llvm:Support",
+        "@llvm-project//mlir:AffineDialect",
+        "@llvm-project//mlir:Analysis",
+        "@llvm-project//mlir:ArithDialect",
+        "@llvm-project//mlir:GPUDialect",
+        "@llvm-project//mlir:IR",
+        "@llvm-project//mlir:SCFDialect",
+        "@llvm-project//mlir:Support",
+    ],
+)
