+++ compiler/src/iree/compiler/Codegen/Common/GPU/ThreadUniformAnalysis.cpp
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/compiler/Codegen/Common/GPU/Analysis/ThreadUniformAnalysis.h"
+#include "iree/compiler/Codegen/Common/GPU/ThreadUniformAnalysis.h"
 #include "iree/compiler/Codegen/Dialect/Codegen/IR/IREECodegenAttrs.h"
 #include "iree/compiler/Codegen/Dialect/GPU/IR/IREEGPUAttrs.h"
 #include "iree/compiler/Dialect/HAL/IR/HALOps.h"
