+++ compiler/src/iree/compiler/Codegen/Common/GPU/ThreadUniformAnalysis.h
@@ -1,8 +1,12 @@
+
 // Copyright 2024 The IREE Authors
 //
 // Licensed under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+#ifndef IREE_COMPILER_CODEGEN_COMMON_GPU_THREADUNIFORMANALYSIS_H_
+#define IREE_COMPILER_CODEGEN_COMMON_GPU_THREADUNIFORMANALYSIS_H_
 
 #include "mlir/Analysis/DataFlow/SparseAnalysis.h"
 
@@ -85,3 +89,5 @@ public:
 
 MLIR_DECLARE_EXPLICIT_TYPE_ID(
     mlir::iree_compiler::dataflow::ThreadUniformLattice)
+
+#endif // IREE_COMPILER_CODEGEN_COMMON_GPU_THREADUNIFORMANALYSIS_H_
